Finished Code need comments
Made it have mobile response.  Put in a print styles and html for print. put in a footer with an hr in it. Changed the JavaScript and made it put at at certain spot when load on mobile. Refined styles. Made the section a flex-container and a flex-box. Made the description list a box and made them look fancy and responsive. Then put in responsive design.